Refactor a lot of the compileClient code into something that can hot-restart the server and reload options from .reactserverrc, routes, and custom webpack options.
