+++ packages/react-server-cli/package.json
@@ -21,6 +21,7 @@
     "babel-runtime": "^6.18.0",
     "body-parser": "^1.15.2",
     "chalk": "^1.1.3",
+    "chokidar": "^1.6.1",
     "chunk-manifest-webpack-plugin": "0.0.1",
     "compression": "~1.6.2",
     "css-loader": "~0.15.4",
@@ -44,7 +45,7 @@
     "sass-loader": "^4.0.0",
     "style-loader": "~0.12.3",
     "webpack": "^1.13.1",
-    "webpack-middleware": "^1.5.1",
+    "webpack-dev-middleware": "^1.9.0",
 	"webpack-hot-middleware": "^2.13.2",
     "webpack-stats-plugin": "^0.1.3",
     "yargs": "~3.15.0"
