+++ packages/react-server-cli/src/defaultOptions.js
@@ -17,6 +17,7 @@ export default {
 	hot: true,
 	minify: false,
 	compileOnly: false,
+	compileOnStartup: true,
 	logLevel: "debug",
 	timingLogLevel: "fast",
 	gaugeLogLevel: "ok",
