+++ packages/react-server-cli/src/run.js
@@ -13,25 +13,6 @@ export default function run(options = {}) {
 	// options neither passed in nor in a config file, use the defaults.
 	options = mergeOptions(defaultOptions, findOptionsInFiles() || {}, options);
 
-	const {
-		routesFile,
-		jsUrl,
-		host,
-		port,
-		httpsOptions,
-	} = options;
-
-	options.routesPath = path.resolve(process.cwd(), routesFile);
-	options.routesDir = path.dirname(options.routesPath);
-
-	try {
-		options.routes = require(options.routesPath);
-	} catch (e) {
-		// Pass. Commands need to check for routes themselves.
-	}
-
-	options.outputUrl = jsUrl || `${httpsOptions ? "https" : "http"}://${host}:${port}/`;
-
 	try {
 		return require("./" + path.join("commands", options.command))(options);
 	} catch (e) {
